Issue #4220: Modified EmptyCatchBlockCheckTest.java and moved its input files to the emptycatchblock subdirectory
